update sdk to 3.0.23
