+++ ros/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <package format="2">
   <name>cepton_ros</name>
-  <version>2.1.0</version>
+  <version>2.1.1</version>
   <description>The cepton ros package</description>
 
   <!-- One maintainer tag required, multiple allowed, one person per tag -->
