+++ ros2/cepton_publisher/include/cepton_publisher.h
@@ -21,7 +21,7 @@ namespace cepton_ros {
 
 // update this when making changes, will display in terminal running publisher
 using String = std::string;
-String VERSION = "v2.1.0";
+String VERSION = "v2.1.1";
 
 enum SensorStatusFlags : uint32_t { SENSOR_TIMED_OUT = 1 << 0 };
