+++ ros2/cepton_publisher/package.xml
@@ -2,7 +2,7 @@
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
   <name>cepton_publisher</name>
-  <version>2.1.0</version>
+  <version>2.1.1</version>
   <description>
     Implements a publisher for Cepton lidar messages. Publishes Cepton info messages, and point cloud data in PointCloud2 format.
   </description>
